requirements_research.txt
# QNTI Research Agent Dependencies

# Core RAG and LLM
llama-index>=0.10.0
llama-index-llms-ollama>=0.1.0
llama-index-embeddings-ollama>=0.1.0

# Web scraping and parsing
requests>=2.31.0
aiohttp>=3.9.0
beautifulsoup4>=4.12.0
feedparser>=6.0.10
markdownify>=0.11.6

# Document processing
PyPDF2>=3.0.1
python-docx>=1.1.0
docx2txt>=0.8

# Alternative RAG (if preferred over LlamaIndex)
# haystack-ai>=2.0.0

# Additional utilities
asyncio-throttle>=1.0.2
python-dateutil>=2.8.2
chardet>=5.2.0

# Optional: Advanced web automation
# playwright>=1.40.0  # For JavaScript-heavy sites
# selenium>=4.15.0    # Alternative web automation 